removed configmap watch from placement_rule
