+++ operators/multiclusterobservability/pkg/config/config.go
@@ -98,11 +98,6 @@ const (
 
 	AllowlistCustomConfigMapName = "observability-metrics-custom-allowlist"
 
-	ManagedClusterLabelConfigMapName         = "observability-managed-cluster-label-names"
-	ManagedClusterLabelBlackListConfigMapKey = "blacklist_label.yaml"
-	ManagedClusterLabelConfigMapKey          = "managed_cluster.yaml"
-	ManagedClusterLabelMetricName            = "managed_cluster_labels"
-
 	ProxyServiceName      = "rbac-query-proxy"
 	ProxyRouteName        = "rbac-query-proxy"
 	ProxyRouteBYOCAName   = "proxy-byo-ca"
@@ -416,26 +411,6 @@ func ReadImageManifestConfigMap(c client.Client, version string) (bool, error) {
 // GetImageManifests...
 func GetImageManifests() map[string]string {
 	return imageManifests
-}
-
-// GetManagedClusterLabelConfigMapKey return the key name for the managedcluster blacklisted labels
-func GetManagedClusterLabelBlackListConfigMapKey() string {
-	return ManagedClusterLabelBlackListConfigMapKey
-}
-
-// GetManagedClusterLabelConfigMapKey return the key name for the managedcluster labels
-func GetManagedClusterLabelConfigMapKey() string {
-	return ManagedClusterLabelConfigMapKey
-}
-
-// GetManagedClusterLabelConfigMapName return the name for the managedcluster labels configmap
-func GetManagedClusterLabelConfigMapName() string {
-	return ManagedClusterLabelConfigMapName
-}
-
-// GetManagedClusterLabelMetricName return the name of the managedcluster label metric
-func GetManagedClusterLabelMetricName() string {
-	return ManagedClusterLabelMetricName
 }
 
 // SetImageManifests sets imageManifests
